+++ app/src/androidTest/kotlin/jp/co/yumemi/android/codecheck/TopActivityTest.kt
@@ -7,16 +7,16 @@ import androidx.test.filters.LargeTest
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import dagger.hilt.android.testing.UninstallModules
 import dagger.hilt.components.SingletonComponent
-import io.ktor.client.*
-import io.ktor.client.engine.android.*
-import io.ktor.util.network.*
+import jp.co.yumemi.android.codecheck.data.remote.Api
+import jp.co.yumemi.android.codecheck.data.remote.ItemRemoteDataSource
+import jp.co.yumemi.android.codecheck.data.remote.MockApi
 import jp.co.yumemi.android.codecheck.di.NetworkModule
 import jp.co.yumemi.android.codecheck.pages.SearchPage
+import jp.co.yumemi.android.codecheck.ui.TopActivity
 import kotlinx.coroutines.Dispatchers
 import org.junit.Rule
 import org.junit.Test
