+++ app/src/androidTest/kotlin/jp/co/yumemi/android/codecheck/pages/SearchPage.kt
@@ -6,7 +6,7 @@ import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.contrib.RecyclerViewActions.actionOnItemAtPosition
 import androidx.test.espresso.matcher.ViewMatchers.*
-import jp.co.yumemi.android.codecheck.CustomAdapter
+import jp.co.yumemi.android.codecheck.ui.search.CustomAdapter
 import jp.co.yumemi.android.codecheck.R
 import jp.co.yumemi.android.codecheck.utils.RecyclerViewUtils.recyclerViewSizeMatcher
 import jp.co.yumemi.android.codecheck.utils.RecyclerViewUtils.withItemViewAtPosition
