--- app/src/main/kotlin/jp/co/yumemi/android/codecheck/Api.kt
@@ -1,5 +1,0 @@
-package jp.co.yumemi.android.codecheck
-
-interface Api {
-    suspend fun fetchRepositories(inputText: String): List<Item>
-}
