+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/data/remote/Api.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.codecheck.data.remote
+
+import jp.co.yumemi.android.codecheck.model.Item
+
+interface Api {
+    suspend fun fetchRepositories(inputText: String): List<Item>
+}
