+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/model/Item.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.codecheck
+package jp.co.yumemi.android.codecheck.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
