+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/ui/TopActivity.kt
@@ -1,10 +1,11 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.codecheck
+package jp.co.yumemi.android.codecheck.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import dagger.hilt.android.AndroidEntryPoint
+import jp.co.yumemi.android.codecheck.R
 
 @AndroidEntryPoint
 class TopActivity : AppCompatActivity(R.layout.activity_top)
