+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/SearchFragment"
-        android:name="jp.co.yumemi.android.codecheck.SearchFragment"
+        android:name="jp.co.yumemi.android.codecheck.ui.search.SearchFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_search">
         <action
@@ -17,12 +17,12 @@
 
     <fragment
         android:id="@+id/RepositoryFragment"
-        android:name="jp.co.yumemi.android.codecheck.RepositoryFragment"
+        android:name="jp.co.yumemi.android.codecheck.ui.repository.RepositoryFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_repository">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.codecheck.Item" />
+            app:argType="jp.co.yumemi.android.codecheck.model.Item" />
     </fragment>
 
 </navigation>
